Change admin_v2 edit room modal from styled components to scss (#1193)
* Delete RoomEdit.styles.ts

* Create RoomEdit.scss

* Update RoomEdit.tsx

* Create RoomEditModal.types.ts

* Create RoomEditModal.scss

* Delete RoomEdit.types.ts

* Delete RoomEdit.tsx

* Delete RoomEdit.scss

* Update index.ts

* Create RoomEditModal.tsx

* Apply suggestions from code review


* Update VenueDetails.tsx

* Delete RoomEditModal.types.ts

* Update RoomEditModal.tsx

* Update RoomEditModal.tsx

* Update RoomEditModal.tsx

* Delete RoomEdit.tsx

* Update RoomEditModal.tsx
