--- src/pages/Admin/Room/Edit/RoomEdit.styles.ts
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div``;
-
-export const ButtonsWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-export const InputWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  margin-bottom: 1.5rem;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  flex-direction: column;
-`;
