--- src/pages/Admin/Room/Edit/RoomEdit.types.ts
@@ -1,9 +1,0 @@
-import { RoomData_v2 } from "types/rooms";
-
-export interface EditRoomProps {
-  isVisible: boolean;
-  onClickOutsideHandler: () => void;
-  room: RoomData_v2;
-  submitHandler: (values: RoomData_v2, index: number) => void;
-  deleteHandler: () => void;
-}
