+++ src/pages/Admin/Room/Edit/RoomEditModal.scss
@@ -1,0 +1,17 @@
+@import "scss/constants.scss";
+
+.room-edit-modal {
+  &__input {
+    display: flex;
+    width: 100%;
+    margin-bottom: 1.5rem;
+    flex-wrap: wrap;
+    align-items: flex-start;
+    flex-direction: column;
+  }
+
+  &__buttons {
+    display: flex;
+    justify-content: space-between;
+  }
+}
