+++ src/pages/Admin/Venue/Details/VenueDetails.tsx
@@ -16,7 +16,7 @@ import { VenueDetailsProps } from "./VenueDetails.types";
 import VenueHero from "components/molecules/VenueHero";
 import Button from "components/atoms/Button";
 import AdminEventModal from "pages/Admin/AdminEventModal";
-import RoomEdit from "pages/Admin/Room/Edit";
+import { RoomEditModal } from "pages/Admin/Room/Edit";
 import RoomModal from "pages/Admin/Room/Modal";
 import RoomCard from "pages/Admin/Room/Card";
 import MapPreview from "pages/Admin/MapPreview";
@@ -256,7 +256,7 @@ const VenueDetails: React.FC<VenueDetailsProps> = ({ venue }) => {
       />
 
       {editingRoom && (
-        <RoomEdit
+        <RoomEditModal
           isVisible={!!editingRoom && !showDeleteModal}
           onClickOutsideHandler={closeEditingModal}
           room={editingRoom}
